Implement submission strategies
